+++ CHANGELOG.rst
@@ -6,6 +6,13 @@ All notable changes to this project will be documented in this file.
 This project adheres to `Semantic Versioning <http://semver.org/>`_.
 
 
+0.1.0
+*****
+* First release.
+* Introduced of four new modules: ``empty``, ``schema``,
+  ``typing_utils`` and ``utils``.
+
+
 [Unreleased]
 ************
 * Empty Python project directory structure.